+++ src/renderer/VideoButton.tsx
@@ -1,15 +1,11 @@
-
 import * as React from 'react';
-import Tab from '@mui/material/Tab';
-import Menu from '@mui/material/Menu';
-import MenuItem from '@mui/material/MenuItem';
-import { categories, videoButtonSx, specToClass, VideoCategory }  from '../main/constants';
-import { getResultText, getFormattedDuration, getInstanceDifficulty } from './rendererutils';
+import { Tab, Menu, MenuItem, Divider } from '@mui/material';
+import { VideoCategory, categories, videoButtonSx, specializationById, dungeonsByMapId, dungeonTimersByMapId }  from 'main/constants';
+import { TimelineSegmentType, ChallengeModeDungeon } from 'main/keystone';
+import { getResultText, getFormattedDuration, getVideoResult, getInstanceDifficulty, getDungeonEncounterById } from './rendererutils';
 import ListItemIcon from '@mui/material/ListItemIcon';
 import Check from '@mui/icons-material/Check';
 import * as Images from './images'
-
-
 
 /**
  * For shorthand referencing. 
@@ -26,8 +22,10 @@ export default function VideoButton(props: any) {
   const video = state.videoState[category][videoIndex];
   const videoPath = video.fullPath;
 
-  const isGoodResult = video.result;
-  const resultText = getResultText(category, isGoodResult);
+  const isGoodResult = getVideoResult(video);
+  // Need to not be const as it will be modified later if the video is of a
+  // Mythic Keystone dungeon.
+  let resultText = getResultText(category, isGoodResult);
   const MMR = video.teamMMR ? ("MMR: " + video.teamMMR) : undefined;
 
   const isProtected = video.protected;
@@ -42,7 +40,7 @@ export default function VideoButton(props: any) {
 
   const playerName = video.playerName;
   const specIcon: string = Images.spec[video.playerSpecID] || Images.spec[0];
-  const playerClass = specToClass[video.playerSpecID];
+  const playerClass = specializationById[video.playerSpecID]?.class ?? '';
   
   const dateDisplay: string = video.isFromToday ? video.time : video.date;
   const dateHoverText = video.date + " " + video.time;
@@ -50,18 +48,27 @@ export default function VideoButton(props: any) {
   // BGs don't log COMBATANT_INFO events so we can't display a lot of stuff
   // that we can for other categories. 
   const isBG = category === VideoCategory.Battlegrounds;
+  const isMythicPlus = (category == VideoCategory.MythicPlus && video.challengeMode !== undefined);
   const isRaid = category === VideoCategory.Raids;
   const videoInstanceDifficulty = isRaid ? getInstanceDifficulty(video.difficultyID) : null;
 
   let buttonImage;
 
-  // TODO, clean this ugly code up. What about M+ eventually?
-  if (isRaid) {
-    buttonImage = Images.raid[video.encounterID];
-  } else if (isBG) {
-    buttonImage = Images.battleground[video.zoneID];
-  } else {
-    buttonImage = Images.arena[video.zoneID];
+  switch (category) {
+    case VideoCategory.Raids:
+      buttonImage = Images.raid[video.encounterID];
+      break;
+
+    case VideoCategory.MythicPlus:
+      buttonImage = Images.dungeon[video.challengeMode.zoneId];
+      break;
+
+    case VideoCategory.Battlegrounds:
+      buttonImage = Images.battleground[video.zoneID];
+      break;
+
+    default:
+      buttonImage = Images.arena[video.zoneID];
   }
 
   /**
@@ -81,7 +88,7 @@ export default function VideoButton(props: any) {
   /**
   * Delete a video.
   */
-  function deleteVideo(filePath: string) {6
+  function deleteVideo(filePath: string) {
     ipc.sendMessage('contextMenu', ['delete', filePath])
     handleCloseMenu();
   };
@@ -102,14 +109,105 @@ export default function VideoButton(props: any) {
     handleCloseMenu();
   };
 
+  /**
+   * Seek the selected video to the specified relative timestamp
+   */
+  const seekVideo = (videoIndex: number, timestamp: number) => {
+    ipc.sendMessage('contextMenu', ['seekVideo', videoIndex, timestamp])
+    handleCloseMenu();
+  };
+
+  /**
+   * Generate the JSX for the timeline segments that are used in the context menu on
+   * the VideoButton.
+   */
+  const renderKeystoneTimelineSegments = (challengeMode: any): any[] => {
+    const timelineSegmentsMenuItems = challengeMode.timelineSegments.map((segment: any) => {
+      let timelineSegmentMenu;
+      let segmentDurationText;
+      const result = Boolean(segment.result);
+
+      // If the metadata for some reason gets a malformed timestamp, let's
+      // not make it break the whole UI but instead silently ignore it for now.
+      try {
+        segmentDurationText = getFormattedDuration(segment.timestamp);
+      } catch (e: any) {
+        console.error(e)
+        return;
+      }
+
+      if (segment.segmentType === TimelineSegmentType.Trash) {
+        timelineSegmentMenu = <div className='segment-type segment-type-trash'>
+          <span>{ segmentDurationText }</span>: Trash
+        </div>
+      } else
+      if (segment.segmentType == TimelineSegmentType.BossEncounter) {
+        timelineSegmentMenu = <div className='segment-entry'>
+          <div className='segment-type segment-type-boss'>
+            <span>{ segmentDurationText }</span>: Boss: { getDungeonEncounterById(segment.encounterId) }
+          </div>
+          <div className={ 'segment-result ' + (result ? 'goodResult' : 'badResult') }>
+            { getResultText(VideoCategory.Raids, result) }
+          </div>
+        </div>
+      }
+
+      return (
+        <MenuItem key={ 'video-segment-' + segment.timestamp } onClick={() => seekVideo(videoIndex, segment.timestamp)}>
+          { timelineSegmentMenu }
+        </MenuItem>
+      );
+    });
+
+    return [
+      <MenuItem key='video-segment-label' disabled>Video Segments</MenuItem>,
+      <Divider key='video-segments-begin' />,
+      ...timelineSegmentsMenuItems,
+      <Divider key='video-segments-end' />,
+    ];
+  };
+
+  const buttonClasses = ['videoButton'];
+  let keystoneTimelineSegments = [];
+
+  if (isMythicPlus) {
+    buttonClasses.push('dungeon')
+
+    // If the video metadata doesn't contain its own alloted time array, we'll use whichever
+    // is current in the constants.
+    const keystoneAllottedTime = video.challengeMode.allottedTime ?? dungeonTimersByMapId[video.challengeMode.mapId];
+
+    // Calculate the upgrade levels of the keystone
+    const keystoneUpgradeLevel = ChallengeModeDungeon.calculateKeystoneUpgradeLevel(
+      keystoneAllottedTime,
+      video.challengeMode.duration
+    );
+
+    if (isGoodResult) {
+      resultText = '+' + keystoneUpgradeLevel;
+    }
+
+    keystoneTimelineSegments = renderKeystoneTimelineSegments(video.challengeMode);
+  }
+
   return (
     <React.Fragment>
       <Tab 
         label={
-          <div id={ videoPath } className='videoButton' style={{ backgroundImage: `url(${buttonImage})`}} onContextMenu={openMenu}>
+          <div id={ videoPath } className={ buttonClasses.join(' ') } style={{ backgroundImage: `url(${buttonImage})`}} onContextMenu={openMenu}>
             <div className='duration'>{ formattedDuration }</div>
-            <div className='encounter'>{ video.encounter }</div>            
-            <div className='zone'>{ video.zone }</div>
+            { isMythicPlus ||
+              <div>
+                <div className='encounter'>{ video.encounter }</div>
+                <div className='zone'>{ video.zone }</div>
+              </div>
+            }
+            { isMythicPlus &&
+              <div>
+                <div className='zone'>{ dungeonsByMapId[video.challengeMode.mapId] }</div>
+                <div className='zone level'>+{ video.challengeMode.level }</div>
+              </div>
+            }
             <div className='time' title={ dateHoverText }>{ dateDisplay }</div>    
             { isRaid && videoInstanceDifficulty &&
               <div className={'instance-difficulty difficulty-' + videoInstanceDifficulty.difficultyId}>
@@ -119,12 +217,12 @@ export default function VideoButton(props: any) {
             { isBG ||
               <div>
                 <div className='specIcon'>
-                  <img src={ specIcon } width='25' height='25'/>
+                  <img src={ specIcon } />
                 </div>  
                 <div className={ playerClass + ' name'}>{ playerName }</div>
               </div>
             }
-            <div className={ resultClass } title={ MMR }>{ resultText }</div>
+            <div className={ 'resultText ' + resultClass } title={ MMR }>{ resultText }</div>
           </div> 
         }
         key={ videoPath }
@@ -137,6 +235,7 @@ export default function VideoButton(props: any) {
         anchorOrigin={{ vertical: 'bottom', horizontal: 'center' }} 
         open={open} onClose={handleCloseMenu} 
         MenuListProps={{'aria-labelledby': 'basic-button'}}>
+        { keystoneTimelineSegments.length > 0 && keystoneTimelineSegments }
         <MenuItem onClick={() => deleteVideo(videoPath)}>Delete</MenuItem>
         <MenuItem onClick={() => saveVideo(videoPath)}> 
           {isProtected &&